From: David Shaw <dshaw@jabberwocky.com> To: bugtraq@securityfocus.com Subject: Key validity bug in GnuPG 1.2.1 and earlier Date: Sat, 3 May 2003 21:35:26 -0400 User-Agent: Mutt/1.5.4i
As part of the development of GnuPG 1.2.2, a bug was discovered in the
key validation code.  This bug causes keys with more than one user ID
to give all user IDs on the key the amount of validity given to the
most-valid key.

This bug does not impact any key with only one user ID.  Photo IDs
("user attribute IDs") do not count as an additional user ID for the
purposes of this bug.

For example, given a key with two user IDs:
   Alice <alice@example.com>
and
   Alice's other address <alice@corp.example.net>

If the encrypting user has a trust path to to the ID
alice@example.com, then this ID is fully valid, and there is no
warning message when encrypting to alice@example.com.

If the encrypting user has either an insufficient or no trust path to
the ID "alice@corp.example.net", then that ID is either not fully
valid, or not valid at all respectively.  There should be a warning
message given when encrypting to this other user ID ("it is not
certain this key belongs to the user named in the user ID / do you
want to encrypt to it anyway?"), but due to the bug, the invalid user
ID is accepted as valid and no warning message is given.

Changed files:
    patch-gnupg-1.2.1-trustfix.txt -> 1.1