- rel 3 - security patch BUg (bugtraq): This bug does not impact any key with only one user ID.  Photo IDs ("user attribute IDs") do not count as an additional user ID for the purposes of this bug.
For example, given a key with two user IDs:
   Alice <alice@example.com>
and
   Alice's other address <alice@corp.example.net>

If the encrypting user has a trust path to to the ID
alice@example.com, then this ID is fully valid, and there is no
warning message when encrypting to alice@example.com.

If the encrypting user has either an insufficient or no trust path to
the ID "alice@corp.example.net", then that ID is either not fully
valid, or not valid at all respectively.  There should be a warning
message given when encrypting to this other user ID ("it is not
certain this key belongs to the user named in the user ID / do you
want to encrypt to it anyway?"), but due to the bug, the invalid user
ID is accepted as valid and no warning message is given.

Changed files:
    gnupg.spec -> 1.73